Extracted empty list composable
